Add repo issue scripts, break out into separate files
