+++ scripts/download-repo-issues/ORNL_repo_total_issues.csv
@@ -1,0 +1,120 @@
+"Repo","Label","Total Issues","Total Open","Total Closed"
+"AMO-Tools-Desktop","Analytics","7","1","6"
+"AMO-Tools-Desktop","Application","248","39","209"
+"AMO-Tools-Desktop","ASAP","61","1","60"
+"AMO-Tools-Desktop","bug","672","21","651"
+"AMO-Tools-Desktop","Calculator","504","41","463"
+"AMO-Tools-Desktop","CO2 Emissions","54","5","49"
+"AMO-Tools-Desktop","Compressed Air","314","28","286"
+"AMO-Tools-Desktop","Compressed Air Inventory","3","3","0"
+"AMO-Tools-Desktop","Data Explorer","105","10","95"
+"AMO-Tools-Desktop","Databases","19","5","14"
+"AMO-Tools-Desktop","devops","1","0","1"
+"AMO-Tools-Desktop","duplicate","35","2","33"
+"AMO-Tools-Desktop","enhancement","443","85","358"
+"AMO-Tools-Desktop","Expert Submitted","19","10","9"
+"AMO-Tools-Desktop","Fan Inventory","3","3","0"
+"AMO-Tools-Desktop","Fans","336","23","313"
+"AMO-Tools-Desktop","github_actions","2","0","2"
+"AMO-Tools-Desktop","Help Text","16","7","9"
+"AMO-Tools-Desktop","help wanted","24","11","13"
+"AMO-Tools-Desktop","important","209","1","208"
+"AMO-Tools-Desktop","Intern To Do","110","46","64"
+"AMO-Tools-Desktop","invalid","0","0","0"
+"AMO-Tools-Desktop","Inventory","5","5","0"
+"AMO-Tools-Desktop","javascript","10","0","10"
+"AMO-Tools-Desktop","low priority","135","41","94"
+"AMO-Tools-Desktop","Motor Inventory","51","3","48"
+"AMO-Tools-Desktop","Motors","25","8","17"
+"AMO-Tools-Desktop","pebkac","2","0","2"
+"AMO-Tools-Desktop","Process Cooling","56","6","50"
+"AMO-Tools-Desktop","Process Heating","602","32","570"
+"AMO-Tools-Desktop","Pump Inventory","5","0","5"
+"AMO-Tools-Desktop","Pumps","338","19","319"
+"AMO-Tools-Desktop","question","63","17","46"
+"AMO-Tools-Desktop","Quick Fix","147","2","145"
+"AMO-Tools-Desktop","Redesign","29","1","28"
+"AMO-Tools-Desktop","Reports","136","5","131"
+"AMO-Tools-Desktop","Stale","37","26","11"
+"AMO-Tools-Desktop","Status Check","83","80","3"
+"AMO-Tools-Desktop","Steam","334","28","306"
+"AMO-Tools-Desktop","suite","7","0","7"
+"AMO-Tools-Desktop","Treasure Hunt","281","31","250"
+"AMO-Tools-Desktop","User Submitted","0","0","0"
+"AMO-Tools-Desktop","Waste Water","79","7","72"
+"AMO-Tools-Desktop","Water","65","31","34"
+"AMO-Tools-Desktop","Web","25","0","25"
+"AMO-Tools-Desktop","WebAssembly","34","0","34"
+"AMO-Tools-Desktop","wontfix","7","3","4"
+"AMO-Tools-Desktop","Total","5741","687","5054"
+"AMO-Tools-Suite","Application","11","4","7"
+"AMO-Tools-Suite","ASAP","8","0","8"
+"AMO-Tools-Suite","bug","45","1","44"
+"AMO-Tools-Suite","Calculator","57","8","49"
+"AMO-Tools-Suite","Compressed Air","25","2","23"
+"AMO-Tools-Suite","Databases","8","5","3"
+"AMO-Tools-Suite","Documentation","3","2","1"
+"AMO-Tools-Suite","duplicate","3","0","3"
+"AMO-Tools-Suite","enhancement","17","4","13"
+"AMO-Tools-Suite","Fans","40","3","37"
+"AMO-Tools-Suite","help wanted","3","2","1"
+"AMO-Tools-Suite","important","19","2","17"
+"AMO-Tools-Suite","Intern To Do","0","0","0"
+"AMO-Tools-Suite","invalid","0","0","0"
+"AMO-Tools-Suite","Inventory","3","3","0"
+"AMO-Tools-Suite","low priority","3","1","2"
+"AMO-Tools-Suite","Motors","5","1","4"
+"AMO-Tools-Suite","Process Cooling","7","1","6"
+"AMO-Tools-Suite","Process Heating","59","6","53"
+"AMO-Tools-Suite","Pumps","31","2","29"
+"AMO-Tools-Suite","question","3","0","3"
+"AMO-Tools-Suite","Quick Fix","3","0","3"
+"AMO-Tools-Suite","Stale","4","3","1"
+"AMO-Tools-Suite","Status Check","4","3","1"
+"AMO-Tools-Suite","Steam","71","6","65"
+"AMO-Tools-Suite","suite","53","1","52"
+"AMO-Tools-Suite","To Do","3","0","3"
+"AMO-Tools-Suite","Treasure Hunt","14","2","12"
+"AMO-Tools-Suite","v.7.0.1-release","0","0","0"
+"AMO-Tools-Suite","v0.2.3-beta","2","0","2"
+"AMO-Tools-Suite","Waste Water","2","0","2"
+"AMO-Tools-Suite","Water","1","0","1"
+"AMO-Tools-Suite","WebAssembly","29","1","28"
+"AMO-Tools-Suite","wontfix","1","0","1"
+"AMO-Tools-Suite","Total","537","63","474"
+"VERIFI","Analysis","114","26","88"
+"VERIFI","bug","157","4","153"
+"VERIFI","CO2 Emissions","75","18","57"
+"VERIFI","Dashboard","24","5","19"
+"VERIFI","Data Input","94","26","68"
+"VERIFI","documentation","3","0","3"
+"VERIFI","duplicate","8","0","8"
+"VERIFI","Electron","6","1","5"
+"VERIFI","enhancement","125","28","97"
+"VERIFI","Example","2","0","2"
+"VERIFI","Expert Submitted","4","4","0"
+"VERIFI","github_actions","1","0","1"
+"VERIFI","good first issue","0","0","0"
+"VERIFI","Graphs","35","11","24"
+"VERIFI","Help Text","34","10","24"
+"VERIFI","help wanted","5","4","1"
+"VERIFI","High Priority","41","4","37"
+"VERIFI","Import/Export","79","13","66"
+"VERIFI","important","4","0","4"
+"VERIFI","Intern To Do","16","5","11"
+"VERIFI","Internationalization","4","4","0"
+"VERIFI","invalid","0","0","0"
+"VERIFI","javascript","6","0","6"
+"VERIFI","Low Priority","23","12","11"
+"VERIFI","Predictors","27","12","15"
+"VERIFI","question","26","10","16"
+"VERIFI","Quick Fix","31","0","31"
+"VERIFI","Regression","14","6","8"
+"VERIFI","Reports","72","19","53"
+"VERIFI","Stale","9","9","0"
+"VERIFI","Status Check","10","10","0"
+"VERIFI","UI/UX","86","18","68"
+"VERIFI","Units","28","1","27"
+"VERIFI","User Submitted","2","2","0"
+"VERIFI","wontfix","1","0","1"
+"VERIFI","Total","1166","262","904"
